+++ packages/sdk/CHANGELOG.md
@@ -9,9 +9,16 @@ Not every commit needs to result in a change to this file (e.g. docs and chore c
 commit that affects the code in a way that consumers might care about should include edits to the
 'Unreleased' section though. Breaking changes should be prefixed with `**BREAKING:**`.
 
-## [Unreleased](https://github.com/airtable/blocks-sdk/compare/v0.0.15...HEAD)
+## [Unreleased](--github-repo-url=https://github.com/airtable/blocks-sdk/compare/v0.0.16...HEAD)
 
-## [0.0.15](https://github.com/airtable/blocks-sdk/compare/v0.0.14...v0.0.15) - 2019-07-03
+## [0.0.16](--github-repo-url=https://github.com/airtable/blocks-sdk/compare/v0.0.15...v0.0.16) - 2019-07-05
+
+-   **BREAKING:** Removed view.visibleFields and view.allFields - use view.selectMetadata instead
+-   Added view.selectMetadata() for querying a view's field order and list of visible fields
+
+## [0.0.15](https://github.com/airtable/blocks-sdk/compare/v0.0.14...@airtable/blocks@0.0.15) - 2019-07-03
+
+No changes.
 
 ## [0.0.14](https://github.com/airtable/blocks-sdk/compare/v0.0.13...v0.0.14) - 2019-07-02